move dto classes in stats directory
